% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R
\name{createModel}
\alias{createModel}
\title{Function to create an MCSimMod Model object}
\usage{
createModel(mName = character(0), mString = character(0), writeTmp = TRUE)
}
\arguments{
\item{mName}{Name of an MCSim model specification file, excluding the file name extension \code{.model}.}

\item{mString}{A character string containing MCSim model specification text.}

\item{writeTmp}{A boolean to write compiled files to tmp (writeTmp = TRUE) or to directory with \code{.model} file (writeTmp = FALSE).}
}
\value{
Model object.
}
\description{
This function creates a \code{Model} object using an MCSim model specification
file or an MCSim model specification string.
}
\examples{
\dontrun{
# Simple model
mod <- createModel("path/to/model")

# Load/compile the model
mod$loadModel()

# Update parameters (P1 and P2)
mod$updateParms(c(P1 = 3, P2 = 1))

# Define times for ODE simulation
times <- seq(from = 0, to = 24, by = 0.1)

# Run the simulation
out <- mod$runModel(times)
}

}
