% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileModel.R
\name{compileModel}
\alias{compileModel}
\title{This function translates a model that has been defined in an MCSim model
(".model") file into the C language (i.e., a ".c" file). It then compiles the
model to create an object code (".o") file and a dynamic linked library
(".dll") file, as well as an R script ("_inits.R") containing several R
functions that can be used for initializing model states and parameters.}
\usage{
compileModel(model_file, c_file, dll_name, dll_file)
}
\arguments{
\item{model_file}{model file name that needs to be compiled}

\item{c_file}{output c_file that is compiled by \code{c_mod}}

\item{dll_name}{dynamic library that has the "derivs" function from a previously compiled model}

\item{dll_file}{Possible previously compiled dll}
}
\description{
Inputs:
mName: String containing the name of the MCSim model. Exclude the file name
suffix ".model". If the function is called from a working directory other
than the one containing the ".model" file, the full path of the ".model"
file should be provided.
}
