% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R
\name{createModel}
\alias{createModel}
\title{MCSimMod class to run a model}
\usage{
createModel(mName = character(0), mString = character(0))
}
\arguments{
\item{mName}{path to model file (without .model extension)}

\item{mString}{string for creating model without .model file}
}
\description{
A class for managing MCSimMod models. This wrapper returns a \code{Model}
object to compile and run MCSim models.
}
\examples{
\dontrun{
# Simple model
mod <- createModel(/path/to/model)

# Load/compile the model
mod$loadModel()

# Update parameters (P1 and P2)
mod$updateParms(c(P1=3, P2=1))

# Define times for ODE simulation
times <- seq(from=0, to=24, by=0.1)

# Run the simulation
out <- mod$runModel(times)

}

}
