% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModel.R
\name{createModel}
\alias{createModel}
\title{MCSimMod class to run a model}
\usage{
createModel(mName = character(0), mString = character(0))
}
\arguments{
\item{mName}{path to model file (without .model extension)}

\item{mString}{string for creating model without .model file}
}
\description{
A class for managing MCSimMod models
}
